refactor: move Scopes, Token and Credential to auth package
